services/diffusion-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
diffusers==0.24.0
transformers==4.35.2
accelerate==0.25.0
torch>=2.0.0
torchvision>=0.15.0
pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
safetensors==0.4.1
